requirements.txt
bokeh==2.0.2
streamlit==0.59.0
pytz==2020.1
requests==2.23.0
pandas==1.0.3
azure_storage==0.36.0
python_dateutil==2.8.1
